rename libquix-prep to libnitrate-seq
