+++ libnitrate-seq/src/core/LibMacro.h
