+++ libnitrate-seq/src/qcall/sys_abort.cc
