+++ libnitrate-seq/src/qcall/sys_emit.cc
