+++ libnitrate-seq/src/qcall/sys_error.cc
