+++ libnitrate-seq/src/qcall/sys_fatal.cc
