+++ libnitrate-seq/src/qcall/sys_fetch.cc
