+++ libnitrate-seq/src/qcall/sys_get.cc
