+++ libnitrate-seq/src/qcall/sys_info.cc
