+++ libnitrate-seq/src/qcall/sys_next.cc
