+++ libnitrate-seq/src/qcall/sys_peek.cc
