+++ libnitrate-seq/src/qcall/sys_random.cc
