+++ libnitrate-seq/src/qcall/sys_starttime.cc
