+++ libnitrate-seq/src/qcall/sys_warn.cc
