+++ snap/snapcraft.yaml
@@ -38,7 +38,7 @@ parts:
         rm $SNAPCRAFT_PART_INSTALL/usr/lib/libquix.a
         rm $SNAPCRAFT_PART_INSTALL/usr/lib/libnitrate-core.a
         rm $SNAPCRAFT_PART_INSTALL/usr/lib/libnitrate-lexer.a
-        rm $SNAPCRAFT_PART_INSTALL/usr/lib/libquix-prep.a
+        rm $SNAPCRAFT_PART_INSTALL/usr/lib/libnitrate-seq.a
         rm $SNAPCRAFT_PART_INSTALL/usr/lib/libquix-parser.a
         rm $SNAPCRAFT_PART_INSTALL/usr/lib/libquix-ir.a
         rm $SNAPCRAFT_PART_INSTALL/usr/lib/libquix-codegen.a
